Fix missing export
